+++ Assets/Plugins/Yuumix/OdinToolkits/Core/Editor/Misc/Versions/OdinToolkitsVersions.cs
@@ -2,6 +2,6 @@ namespace Yuumix.OdinToolkits.Core.Editor
 {
     public static class OdinToolkitsVersions
     {
-        public const string VERSION = "Odin Toolkits v0.3.2 - alpha";
+        public const string VERSION = "Odin Toolkits v0.3.3 - alpha";
     }
 }