+++ Assets/Plugins/Yuumix/OdinToolkits/Modules/Editor/DirectoryTreeGen/DirectoryTreeGenToolSO.cs
@@ -4,7 +4,6 @@ using Sirenix.Serialization;
 using UnityEngine;
 using Yuumix.OdinToolkits.Core;
 using Yuumix.OdinToolkits.Core.Editor;
-using Yuumix.OdinToolkits.Modules.Editor;
 
 namespace Yuumix.OdinToolkits.Modules.Editor
 {