+++ Assets/Plugins/Yuumix/OdinToolkits/Modules/Editor/ScriptDocGen/Misc/DocCategory.cs
@@ -1,8 +1,0 @@
-namespace Yuumix.OdinToolkits.Modules.Editor
-{
-    public enum DocCategory
-    {
-        ScriptingAPI,
-        CompleteReferences
-    }
-}