+++ Assets/Plugins/Yuumix/OdinToolkits/Modules/ScriptDocGen/Editor/Misc/DocCategory.cs
@@ -1,0 +1,8 @@
+namespace Yuumix.OdinToolkits.Modules.ScriptDocGen.Editor
+{
+    public enum DocCategory
+    {
+        ScriptingAPI,
+        CompleteReferences
+    }
+}