+++ Assets/Plugins/Yuumix/OdinToolkits/Modules/ScriptDocGen/Editor/Misc/ToolUsageMode.cs
@@ -1,0 +1,10 @@
+namespace Yuumix.OdinToolkits.Modules.ScriptDocGen.Editor
+{
+    public enum ToolUsageMode
+    {
+        SingleType,
+        MultipleTypes,
+        SingleAssembly,
+        MultipleAssemblies
+    }
+}