+++ Assets/Plugins/Yuumix/CHANGELOG.md
@@ -1,14 +1,20 @@
 # CHANGELOG
 
+这个项目的所有显著变化都将记录在这个文件中。
+
+格式基于[Keep a Changelog](https://keepachangelog.com/en/1.1.0/)，以及这个项目 `尽量` 遵循[Semantic Versioning](https://semver.org/spec/v2.0.0.html)。
+
+---
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
----
+## [0.3.3 - alpha] - 2025-08-17
 
-这个项目的所有显著变化都将记录在这个文件中。
+- Alpha 阶段不保证兼容性，内容随时可能修改。
 
-格式基于[Keep a Changelog](https://keepachangelog.com/en/1.1.0/)，以及这个项目 `尽量` 遵循[Semantic Versioning](https://semver.org/spec/v2.0.0.html)。
+- 目前 `v0.3.3 - alpha` 已经可以支持基础功能运行，具体查看 [Odin Toolkits For Unity](https://github.com/yuumixcode/OdinToolkits-For-Unity)
 
 ## [0.1.3.beta] - 2025-05-21